model_server/requirements.txt
fastapi
uvicorn
torch
transformers
pillow
numpy